+++ src/models/index.ts
@@ -21124,6 +21124,12 @@ export interface Trait {
      */
     name: string;
     /**
+     * What the account is able to use this trait for. Contact support to change access.
+     * @type {Array<TraitPermission>}
+     * @memberof Trait
+     */
+    permissions?: Array<TraitPermission>;
+    /**
      * 
      * @type {ResourceType}
      * @memberof Trait
@@ -21141,6 +21147,12 @@ export interface Trait {
      * @memberof Trait
      */
     status: ResourceStatus;
+    /**
+     * 
+     * @type {TraitTier}
+     * @memberof Trait
+     */
+    tier?: TraitTier;
     /**
      * 
      * @type {PrimitiveDataType}
@@ -21174,6 +21186,7 @@ export enum TraitCategory {
     FigProperty = 'fig/property',
     FigReachability = 'fig/reachability',
     FigSociety = 'fig/society',
+    FigFinancial = 'fig/financial',
     UserDefined = 'user_defined'
 }
 /**
@@ -21474,6 +21487,15 @@ export interface TraitMergePatch {
     unit?: string | null;
 }
 /**
+ * Where the account can use the trait.
+ * @export
+ * @enum {string}
+ */
+export enum TraitPermission {
+    AddToScopePayload = 'add_to_scope_payload',
+    DefineCohort = 'define_cohort'
+}
+/**
  * (Parameters used to POST a new value of the `Trait` type.)
  * 
  * A fact about a person. 
@@ -21580,6 +21602,15 @@ export enum TraitStatisticalType {
     Multicategorical = 'multicategorical',
     Ordinal = 'ordinal',
     Nominal = 'nominal'
+}
+/**
+ * A broad category describing the flavor of a trait.
+ * @export
+ * @enum {string}
+ */
+export enum TraitTier {
+    Standard = 'standard',
+    Premium = 'premium'
 }
 /**
  * Information about files previously uploaded as Datasets.