+++ CHANGELOG.md
@@ -10,6 +10,13 @@ Until we reach API 1.0, the following special rules apply:
 1. If you **add a feature** or **fix a bug**, please bump the version from **0.x.y** to **0.x.(y+1)**.
 2. If you **make a breaking change**, please bump the version from **0.x.y** to **0.(x+1).0**.
 
+## [0.10.30] - 2024-09-04
+
+### Added
+
+- `GET /traits.csv` to return information about all traits the user has access to, in csv form.
+- `GET /traits` and `GET /traits/{id}` added `tier` and `permissions` for each trait.
+
 ## [0.10.29] - 2024-08-30
 
 ### Added